src/requirements.txt
Django==2.1.5
django-bootstrap4==0.0.7
django-location-field==2.0.6
django-map-widgets==0.2.1
Pillow==5.4.1
pytz==2018.9
six==1.12.0
